+++ SubModule.cs
@@ -138,6 +138,7 @@ namespace Byzantium1071
             B1071_GovernanceBehavior.Instance = null;
             B1071_DevastationBehavior.Instance = null;
             B1071_CastleRecruitmentBehavior.Instance = null;
+            B1071_VillageInvestmentBehavior.Instance = null;
             B1071_DevastationBehavior.ResetDynamicPatchFlag();
             B1071_OverlayController.Reset();
 
@@ -159,6 +160,7 @@ namespace Byzantium1071
             B1071_GovernanceBehavior.Instance = null;
             B1071_DevastationBehavior.Instance = null;
             B1071_CastleRecruitmentBehavior.Instance = null;
+            B1071_VillageInvestmentBehavior.Instance = null;
             B1071_OverlayController.Reset();
             _exceptionCounts.Clear();
         }
@@ -194,6 +196,7 @@ namespace Byzantium1071
                 starter.AddBehavior(new Byzantium1071.Campaign.Behaviors.B1071_GovernanceBehavior());
                 starter.AddBehavior(new Byzantium1071.Campaign.Behaviors.B1071_DevastationBehavior());
                 starter.AddBehavior(new Byzantium1071.Campaign.Behaviors.B1071_CastleRecruitmentBehavior());
+                starter.AddBehavior(new Byzantium1071.Campaign.Behaviors.B1071_VillageInvestmentBehavior());
                 // Volunteer model is now a Harmony Postfix (B1071_ManpowerVolunteerPatch)
                 // instead of AddModel, for compatibility with mods that replace VolunteerModel.
                 starter.AddModel(new Byzantium1071.Campaign.Models.B1071_ManpowerMilitiaModel());